move modules that don't implement visualizers in view_spatial/src/visualizers into subfolder
