+++ crates/re_space_view_spatial/src/visualizers/depth_images.rs
@@ -27,10 +27,7 @@ use crate::{
     PickableImageRect, SpatialSpaceView2D, SpatialSpaceView3D,
 };
 
-use super::{
-    tensor_to_textured_rect, textured_rect_utils::bounding_box_for_textured_rect,
-    SpatialViewVisualizerData,
-};
+use super::{bounding_box_for_textured_rect, tensor_to_textured_rect, SpatialViewVisualizerData};
 
 pub struct DepthImageVisualizer {
     pub data: SpatialViewVisualizerData,
