+++ crates/re_space_view_spatial/src/visualizers/utilities/entity_iterator.rs
