+++ crates/re_space_view_spatial/src/visualizers/utilities/labels.rs
@@ -1,0 +1,28 @@
+use re_entity_db::InstancePathHash;
+
+#[derive(Clone)]
+pub enum UiLabelTarget {
+    /// Labels a given rect (in scene coordinates)
+    Rect(egui::Rect),
+
+    /// Labels a given point (in scene coordinates)
+    Point2D(egui::Pos2),
+
+    /// A point in space.
+    Position3D(glam::Vec3),
+}
+
+#[derive(Clone)]
+pub struct UiLabel {
+    pub text: String,
+    pub color: egui::Color32,
+
+    /// The shape/position being labeled.
+    pub target: UiLabelTarget,
+
+    /// What is hovered if this label is hovered.
+    pub labeled_instance: InstancePathHash,
+}
+
+/// Maximum number of labels after which we stop displaying labels for that entity all together.
+pub const MAX_NUM_LABELS_PER_ENTITY: usize = 10;
