+++ crates/re_space_view_spatial/src/visualizers/utilities/textured_rect.rs
